src/framework/mlt_geometry.c src/modules/core/transition_composite.c src/modules/sdl/consumer_sdl.c src/modules/sdl/consumer_sdl_still.c + replaced floats with doubles (attempt to avoid rounding errors?)
src/framework/mlt_tractor.c
+ corrections for fx_cuts (allows animated fx)


git-svn-id: https://mlt.svn.sourceforge.net/svnroot/mlt/trunk/mlt@739 d19143bc-622f-0410-bfdd-b5b2a6649095

+++ src/framework/mlt_geometry.c
@@ -98,9 +98,9 @@ static void mlt_geometry_virtual_refresh( mlt_geometry this )
 					geometry_item prev = current->prev;
 					geometry_item next = current->next;
 
-					float prev_value = 0;
-					float next_value = 0;
-					float value = 0;
+					double prev_value = 0;
+					double next_value = 0;
+					double value = 0;
 
 					while( prev != NULL && !prev->data.f[ i ] ) prev = prev->prev;
 					while( next != NULL && !next->data.f[ i ] ) next = next->next;
